+++ src/finam/interfaces.py
@@ -138,6 +138,14 @@ class IComponent(ABC):
     def status(self):
         """The component's current status."""
 
+    @property
+    @abstractmethod
+    def metadata(self):
+        """
+        The component's meta data.
+        Will only be called after the connect phase of the composition.
+        """
+
 
 class ITimeComponent(IComponent, ABC):
     """Interface for components with a time step."""
@@ -452,6 +460,14 @@ class IOutput(ABC):
 
 class IAdapter(IInput, IOutput, ABC):
     """Interface for adapters."""
+
+    @property
+    @abstractmethod
+    def metadata(self):
+        """
+        The adapter's meta data.
+        Will only be called after the connect phase of the composition.
+        """
 
 
 class NoBranchAdapter: