+++ src/finam/sdk/component.py
@@ -219,6 +219,16 @@ class Component(IComponent, Loggable, ABC):
         return self._name
 
     @property
+    def metadata(self):
+        """
+        The component's meta data.
+        Will only be called after the connect phase of the composition.
+
+        Returns an empty ``dict`` unless overwritten in component implementation.
+        """
+        return {}
+
+    @property
     def logger_name(self):
         """Logger name derived from base logger name and class name."""
         base_logger = logging.getLogger(self.base_logger_name)